+++ fio-stl/405 tls13.h
@@ -1612,7 +1612,7 @@ FIO_SFUNC void fio___tls13_finish(int fd, void *tls_ctx) {
 
     if (enc_len > 0) {
       /* Best effort send, ignore errors */
-      int ignr_ = fio_sock_write(fd, (char *)alert, (size_t)enc_len);
+      ssize_t ignr_ = fio_sock_write(fd, (char *)alert, (size_t)enc_len);
       (void)ignr_;
     }
   }
